pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pm-party-client"
version = "0.1.0"
description = "Python SDK for Project Manager third-party authentication (Ed25519 + JWT + GraphQL)"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Shahriyar Tarnasi", email = "shahryar.tarnasi@gmail.com" },
]
keywords = ["graphql", "jwt", "ed25519", "authentication", "sdk"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
]

[tool.hatch.build.targets.wheel]
packages = ["src/pm_party_client"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
